move heading tests to automated pipeline feature (#35933)
