+++ tests/rendering/rest.js
@@ -151,36 +151,6 @@ describe('REST references docs', () => {
       }
     }
   })
-
-  describe('headings', () => {
-    test('rest pages do not render any headings with duplicate text', async () => {
-      const $ = await getDOM('/en/rest/actions/artifacts')
-      const headingText = $('body')
-        .find('h2, h3, h4, h5, h6')
-        .map((i, el) => $(el).text())
-        .get()
-        .sort()
-
-      const dupes = headingText.filter((item, index) => headingText.indexOf(item) !== index)
-
-      const message = `The following duplicate heading texts were found: ${dupes.join(', ')}`
-      expect(dupes.length, message).toBe(0)
-    })
-
-    test('rest pages do not render any headings with duplicate ids', async () => {
-      const $ = await getDOM('/en/rest/actions/artifacts')
-      const headingIDs = $('body')
-        .find('h2, h3, h4, h5, h6')
-        .map((i, el) => $(el).attr('id'))
-        .get()
-        .sort()
-
-      const dupes = headingIDs.filter((item, index) => headingIDs.indexOf(item) !== index)
-
-      const message = `The following duplicate heading IDs were found: ${dupes.join(', ')}`
-      expect(dupes.length, message).toBe(0)
-    })
-  })
 })
 
 function formatErrors(differences) {
